+++ packages/react/Popover/src/popover.spec.tsx
@@ -40,7 +40,7 @@ describe("GIVEN <Popover />", () => {
       const arrow = screen.getByTestId("arrow-element");
       expect(popover).toBeDefined();
 
-      expect(popover.style.top).toEqual("-8px");
+      expect(popover.style.top).toEqual("-10px");
       expect(popover.style.left).toEqual("0px");
       expect(popover.style.position).toEqual("fixed");
 
@@ -58,7 +58,7 @@ describe("GIVEN <Popover />", () => {
       const arrow = screen.getByTestId("arrow-element");
       expect(popover).toBeDefined();
 
-      expect(popover.style.top).toEqual("8px");
+      expect(popover.style.top).toEqual("10px");
       expect(popover.style.left).toEqual("0px");
       expect(popover.style.position).toEqual("fixed");
 
@@ -77,7 +77,7 @@ describe("GIVEN <Popover />", () => {
       expect(popover).toBeDefined();
 
       expect(popover.style.top).toEqual("0px");
-      expect(popover.style.left).toEqual("-8px");
+      expect(popover.style.left).toEqual("-10px");
       expect(popover.style.position).toEqual("fixed");
 
       expect(arrow.style.top).toEqual("0px");
@@ -95,7 +95,7 @@ describe("GIVEN <Popover />", () => {
       expect(popover).toBeDefined();
 
       expect(popover.style.top).toEqual("0px");
-      expect(popover.style.left).toEqual("8px");
+      expect(popover.style.left).toEqual("10px");
       expect(popover.style.position).toEqual("fixed");
 
       expect(arrow.style.top).toEqual("0px");